Add egui_kittest snapshot tests for visual regression testing
Split InfernoApp from main.rs into lib.rs so integration tests can
import it. Add load_and_analyze() for synchronous test setup. Six
snapshot tests cover summary, detail, and track map views in both
dark and light mode. Document snapshot workflow in CLAUDE.md.
