+++ .github/workflows/ci-rust.yml
@@ -76,3 +76,39 @@ jobs:
 
       - name: Run integration tests
         run: cargo test --test integration_xrz
+
+  test-snapshots:
+    name: Snapshot Tests
+    runs-on: ubuntu-latest
+    continue-on-error: true
+    defaults:
+      run:
+        working-directory: desktop_app_rs
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Install system dependencies
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y libxcb-render0-dev libxcb-shape0-dev libxcb-xfixes0-dev libxkbcommon-dev libssl-dev mesa-vulkan-drivers
+
+      - name: Cache Cargo
+        uses: actions/cache@v4
+        with:
+          path: |
+            ~/.cargo/registry
+            ~/.cargo/git
+            desktop_app_rs/target
+          key: snapshots-${{ runner.os }}-${{ hashFiles('desktop_app_rs/Cargo.lock') }}
+          restore-keys: snapshots-${{ runner.os }}-
+
+      - name: Run snapshot tests
+        run: cargo test --test snapshot_tests
+
+      - name: Upload snapshot artifacts on failure
+        if: failure()
+        uses: actions/upload-artifact@v4
+        with:
+          name: snapshot-test-results
+          path: desktop_app_rs/crates/inferno-app/tests/snapshots/
+          retention-days: 7
